L3 Removal, XBar and Cache Hierarchy swapped
