+++ SiFiveP550/P550CacheHierarchy.py
@@ -1,0 +1,169 @@
+# Copyright (c) 2015 Jason Power
+# All rights reserved.
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are
+# met: redistributions of source code must retain the above copyright
+# notice, this list of conditions and the following disclaimer;
+# redistributions in binary form must reproduce the above copyright
+# notice, this list of conditions and the following disclaimer in the
+# documentation and/or other materials provided with the distribution;
+# neither the name of the copyright holders nor the names of its
+# contributors may be used to endorse or promote products derived from
+# this software without specific prior written permission.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+# "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+# LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+# A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
+# OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
+# SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
+# LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+# DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
+# THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+# (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+""" Caches with options for a simple gem5 configuration script
+
+This file contains crossbars for L2 and L3 to be used in the simple
+gem5 configuration file.
+"""
+
+import m5
+import os
+
+m5.util.addToPath(os.environ['GEM_CONFIGS'])
+
+from common import SimpleOpts
+
+from gem5.components.boards.abstract_board import *
+from gem5.components.cachehierarchies.abstract_cache_hierarchy import AbstractCacheHierarchy
+from gem5.components.cachehierarchies.abstract_two_level_cache_hierarchy import AbstractTwoLevelCacheHierarchy
+from gem5.components.cachehierarchies.classic.abstract_classic_cache_hierarchy import AbstractClassicCacheHierarchy
+
+from P550Caches import *
+
+class P550CacheHierarchy(
+    AbstractClassicCacheHierarchy, AbstractTwoLevelCacheHierarchy
+):
+    """
+    Cache hierarchy for the SiFiveP550 processor.
+
+    2-level cache hierarchy:
+    - L1 I/D cache
+    - Private L2
+
+    Also incorporates an MMU cache, although details around the actual
+    MMU used in the P550 are scarce, it is a Sv39 unit.
+
+    Templated from gem5/src/python/gem5/components/cachehierarchies/clasic/
+    private_l1_shared_l2_cache_hierarchy.py
+    """
+
+    def __init__(
+        self,
+        system_bus: Optional[BaseXBar] = None,
+        l1i_size: str = "32KiB",
+        l1d_size: str = "32KiB",
+        l2_size: str = "256KiB",
+        l1i_assoc: int = 4,
+        l1d_assoc: int = 4,
+        l2_assoc: int = 8,
+    ) -> None:
+        """
+        Instantiation of the cache hierarchy.
+        """
+
+        AbstractClassicCacheHierarchy.__init__(self=self)
+        AbstractTwoLevelCacheHierarchy.__init__(
+            self=self,
+            l1i_size=l1i_size,
+            l1i_assoc=l1i_assoc,
+            l1d_size=l1d_size,
+            l1d_assoc=l1d_assoc,
+            l2_size=l2_size,
+            l2_assoc=l2_assoc,
+        )
+
+        if system_bus is None:
+            self.system_bus = SystemXBar(width=64)
+
+    @overrides(AbstractCacheHierarchy)
+    def incorporate_cache(self, board: AbstractBoard) -> None:
+        """
+        Creates the self-contained cache hierarchy for use in the system.
+
+        This function creates an instance of the L1 and L2 caches for
+        each core in the processor, connecting them together with the
+        L2XBar, then creates an instance of the L3 cache, connecting
+        all L2 caches to the L3 through the L3XBar. Finally, it connects
+        the entire cache hierarchy to the CPU by the system port
+        """
+
+        # Connect to the system
+        board.connect_system_port(self.system_bus.cpu_side_ports)
+
+        for _, port in board.get_memory().get_mem_ports():
+            self.system_bus.mem_side_ports = port
+
+        # Create the caches
+        cores = board.get_processor().get_num_cores()
+        self.l1icaches = [
+            L1ICache(
+                size=self._l1i_size,
+                assoc=self._l1i_assoc
+            )
+            for _ in range(cores)
+        ]
+
+        self.l1dcaches = [
+            L1DCache(
+                size=self._l1d_size,
+                assoc=self._l1d_assoc
+            )
+            for _ in range(cores)
+        ]
+
+        self.l2bus = [P550L2XBar() for _ in range(cores)]
+
+        self.l2caches = [
+            L2Cache(
+                size=self._l2_size,
+                assoc=self._l2_assoc
+            )
+            for _ in range(cores)
+        ]
+
+        # ITLB Page walk caches
+        self.iptw_caches = [
+            MMUCache(size="8KiB", writeback_clean=False)
+            for _ in range(board.get_processor().get_num_cores())
+        ]
+        # DTLB Page walk caches
+        self.dptw_caches = [
+            MMUCache(size="8KiB", writeback_clean=False)
+            for _ in range(board.get_processor().get_num_cores())
+        ]
+
+        for i, cpu in enumerate(board.get_processor().get_cores()):
+            # L1 Caches
+            cpu.connect_icache(self.l1icaches[i].cpu_side)
+            cpu.connect_dcache(self.l1dcaches[i].cpu_side)
+
+            self.l1icaches[i].mem_side = self.l2bus[i].cpu_side_ports
+            self.l1dcaches[i].mem_side = self.l2bus[i].cpu_side_ports
+            self.iptw_caches[i].mem_side = self.l2bus[i].cpu_side_ports
+            self.dptw_caches[i].mem_side = self.l2bus[i].cpu_side_ports
+
+            # MMU Caches
+            cpu.connect_walker_ports(
+                self.iptw_caches[i].cpu_side, self.dptw_caches[i].cpu_side
+            )
+
+            # L2 cache connection
+            self.l2caches[i].cpu_side = self.l2bus[i].mem_side_ports
+            self.l2caches[i].mem_side = self.system_bus.cpu_side_ports
+
+            cpu.connect_interrupt()
+
